More error conversions.
Note that *_TRESPASSPATH is meant for situations where something SHOULDN'T ever execute, but where you worry that it might anyways. Currently only for the end of functions that are meant to exit elsewhere, but may prove more useful.

Also, *_TRESPASSPATH may need to be "moved" further down the list, since the group right before it is incomplete...