Note in TODO about that likely symbol lexing error.
Add skeleton code for one of the closing brackets- we can worry about the other cases, and the matching openers, later.
Convert more code to use enqueue_returns() instead of directly pushing retframes.
Add more basic stack functions- inclusive-or for 2, 3, and 4 args, and exclusive-or for two args.
Fill in those forgotten instruction counts for enqueue_returns()...
Add detection code for the closing brackets.
So, the big preprocessor opening-bracket checker was mistakenly using "and" instead of "or"- a token would have needed to SIMULTANEOUSLY be all three types to match. So, it's been fixed to use inclusive-or.

Signed-off-by: Jared Maddox <absinthdraco@gmail.com>