Fix up some includes (e.g. missing semicolons, missing def of dup() ). Add needed stack operations (e.g. drop(), conditionals). Add some stack delta descriptions.
Add a generic reporter for if we encounter in illegal token somewhere.
Add a comma detector.
Add the def for errs{} into basic_toktools.c

Improve the argument name hanmdling for the up-parenthese packager. Note that it now is only right for macro definitions and not invocations: need to fix that.

Signed-off-by: Jared Maddox <absinthdraco@gmail.com>