customprint_signature in inner_err.h needs to be a typedef, not an instance: fix that.
In err.c:
The psl_dshort and psl_short cases in intprint() and unsignedprint() need different typing in their va_arg() calls.
The vals usage in va_args() in ptrprint() needs to be dereferenced.
The definition of customprint_signature as a function pointer in ptrprint() should be deleted.
msg_inner() should take a va_list by pointer, not by value.
Check for a null va_list* in msg_inner();
Add a semicolon to terminate the declaration of scratch in msg_inner().
Set the argument to dynchar_pascalarray_build() in msg_inner() to something small but reasonable.
The loop in msg_inner() needs to dereference txt.
"spec" in msg_inner() is a direct referenc e, not a pointer, so use the dot operator.
All the uses of the "vals" va_list* in the switch in msg_inner() need to be by value, not by taking the pointer.
msg_inner() doesn't have a "text" name visible to it, remove the 'e' to make it txt.
The LIB4_MONAD_EITHER_BODYMATCH() near the end of msg_inner() should just throw away the "success" value instead of saving it, since we never use it.
msg_interface() doesn't have a "vals" va_list visible to it, swap with "args".
msg_interface() needs to give a va_list POINTER to msg_inner(), not a va_list value: change both instances.

Signed-off-by: Jam555 <3349478+jam555@users.noreply.github.com>