Add a test file for the source.h/.c files, and prepare them all for compilation.
source.c:
discard_source() can't directly return dealloc()'s return, add handler code to convert.
Remove all but the last refed_pstr_incrrefs() call from charin() to get the right reference counting.
Alter "source file reporting" in charin() to avoid null dereferencing, and clear it to null if there's no source left.
Alter the "if EOF" that follows the end of the while loop in charin() to return LIB4_RESULT_FAILURE_EOF instead of *_UNDIFFERENTIATED: the EOF macro didn't exist when this was first written.
Alter the *_??? return in charin() to use *_LOGICFAULT: it was added just for this.

source.h:
Alter the "lib4_" stuff to "libandria4_" for both cases.

Signed-off-by: Jam555 <3349478+jam555@users.noreply.github.com>